feat: refactor process attachment display editor and introduce attachment set select component
